requirement.txt
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.9
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psutil==5.8.0
python-dotenv==0.15.0
six==1.15.0
Werkzeug==1.0.1
